test: add Generate Hierarchy Actions test
